% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitytrails-package.R
\docType{package}
\name{securitytrails}
\alias{securitytrails}
\alias{securitytrails-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/securitytrails}
\item BugReports: \url{https://gitlab.com/hrbrmstr/securitytrails/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
