% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-associated-domains.R
\name{st_assoc_domains}
\alias{st_assoc_domains}
\title{Find associated domains}
\usage{
st_assoc_domains(domain, trails_key = trails_api_key())
}
\arguments{
\item{domain}{domain to get}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Find all domains that are related to a domain you input
}
\examples{
st_assoc_domains("oracle.com")
}
\seealso{
Other Domain details: \code{\link{st_domain}},
  \code{\link{st_subdomains}}, \code{\link{st_tags}},
  \code{\link{st_whois}}
}
\concept{Domain details}
