% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-history.R
\name{st_dns_history}
\alias{st_dns_history}
\title{DNS history by record type}
\usage{
st_dns_history(domain, type = c("a", "aaaa", "mx", "ns", "soa", "txt"),
  page = 1L, trails_key = trails_api_key())
}
\arguments{
\item{domain}{domain to get}

\item{type}{one of \code{a}, \code{aaaa}, \code{mx}, \code{ns}, \code{soa} or \code{txt}}

\item{page}{page # of the return results}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Lists out specific historical information about the given hostname parameter
}
\note{
use the \code{pages} list element to determine if you need to call this function
again to retrieve more results from the result set.
}
\examples{
st_dns_history("oracle.com")
}
\seealso{
Other History functions: \code{\link{st_whois_history}}
}
\concept{History functions}
