% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-get-domain.R
\name{st_domain}
\alias{st_domain}
\title{Get Domain}
\usage{
st_domain(domain, trails_key = trails_api_key())
}
\arguments{
\item{domain}{domain to get}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Returns the current data about the given domain. In addition to the current
data, you also get the current statistics associated with a particular record.
For example, for a records you'll get how many other domains have the same IP.
}
\examples{
st_domain("oracle.com")
}
\seealso{
Other Domain details: \code{\link{st_assoc_domains}},
  \code{\link{st_subdomains}}, \code{\link{st_tags}},
  \code{\link{st_whois}}
}
\concept{Domain details}
