% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-explore-ips.R
\name{st_explore_ips}
\alias{st_explore_ips}
\title{Explore IPs}
\usage{
st_explore_ips(ip_address, trails_key = trails_api_key())
}
\arguments{
\item{ip_address}{Starting IP address (optionally with CIDR subnet mask)}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Returns the neighbors in any given IP level range and essentially allows you to explore closeby IP addresses.
}
\examples{
st_explore_ips("8.8.8.8")
}
\concept{IP functions}
