% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-ping.R
\name{st_ping}
\alias{st_ping}
\title{Test Access To the SecurityTrails API}
\usage{
st_ping(trails_key = trails_api_key())
}
\arguments{
\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\description{
Test Access To the SecurityTrails API
}
\examples{
st_ping()
}
