% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-subdomains.R
\name{st_subdomains}
\alias{st_subdomains}
\title{List Subdomains}
\usage{
st_subdomains(domain, trails_key = trails_api_key())
}
\arguments{
\item{domain}{domain to get}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Returns subdomains for a given hostname
}
\examples{
st_subdomains("oracle.com")
}
\seealso{
Other Domain details: \code{\link{st_assoc_domains}},
  \code{\link{st_domain}}, \code{\link{st_tags}},
  \code{\link{st_whois}}
}
\concept{Domain details}
