% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-tags.R
\name{st_tags}
\alias{st_tags}
\title{List Tags}
\usage{
st_tags(domain, trails_key = trails_api_key())
}
\arguments{
\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}

\item{host}{hostname}
}
\value{
a list
}
\description{
Returns tags for a given hostname
}
\examples{
st_tags("oracle.com")
}
\seealso{
Other Domain details: \code{\link{st_assoc_domains}},
  \code{\link{st_domain}}, \code{\link{st_subdomains}},
  \code{\link{st_whois}}
}
\concept{Domain details}
