% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-whois.R
\name{st_whois}
\alias{st_whois}
\title{Get WHOIS}
\usage{
st_whois(domain, trails_key = trails_api_key())
}
\arguments{
\item{domain}{domain to get}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Returns the current WHOIS data about a given domain with the stats merged together
}
\examples{
st_whois("oracle.com")
}
\seealso{
Other Domain details: \code{\link{st_assoc_domains}},
  \code{\link{st_domain}}, \code{\link{st_subdomains}},
  \code{\link{st_tags}}
}
\concept{Domain details}
