% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-whois-history.R
\name{st_whois_history}
\alias{st_whois_history}
\title{WHOIS history by domain}
\usage{
st_whois_history(domain, page = 1L, trails_key = trails_api_key())
}
\arguments{
\item{domain}{domain to get}

\item{page}{page # of the return results}

\item{trails_key}{your \href{https://docs.securitytrails.com/docs/authentication}{SecurityTrails API Key}.
See \code{\link[=trails_api_key]{trails_api_key()}}.}
}
\value{
a list
}
\description{
Lists out specific historical information about the given hostname parameter
}
\note{
use the \code{pages} list element to determine if you need to call this function
again to retrieve more results from the result set.
}
\examples{
st_whois_history("oracle.com")
}
\seealso{
Other History functions: \code{\link{st_dns_history}}
}
\concept{History functions}
