% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-api-key.R
\name{trails_api_key}
\alias{trails_api_key}
\title{Get or set SECURITYTRAILS_API_KEY value}
\usage{
trails_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new SecurityTrails API key for the current environment?}
}
\value{
atomic character vector containing the SecurityTrails API key
}
\description{
The API wrapper functions in this package all rely on a SecurityTrails API
key residing in the environment variable \code{SECURITYTRAILS_API_KEY}. The
easiest way to accomplish this is to set it in the `~.Renviron`` file in your
home directory.
}
