% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitytrails-package.R
\docType{package}
\name{securitytrails}
\alias{securitytrails}
\alias{securitytrails-package}
\title{Tools to Query the 'SecurityTrails' 'API'}
\description{
'SecurityTrails' (<securitytrails.com>) offers an 'API' to query
information about domains, 'IP' addresses, hostnames and more. Methods are
provided to query 'SecurityTrails' 'API' endpoints.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/securitytrails}
\item BugReports: \url{https://gitlab.com/hrbrmstr/securitytrails/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
